Rename jsx and scss files
