+++ src/components/Ellipsis/Ellipsis.module.scss
