+++ src/components/Navigator/Navigator.module.scss
