+++ src/components/Player/Player.module.scss
