+++ src/components/ProgressBar/ProgressBar.module.scss
