+++ src/components/ResizableContainer/ResizableContainer.module.scss
