+++ src/components/VisualizationViewer/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { BaseComponent } from 'components';
 import { actions } from 'reducers';
-import styles from './stylesheet.module.scss';
+import styles from './VisualizationViewer.module.scss';
 import * as TracerClasses from 'core/tracers';
 import * as LayoutClasses from 'core/layouts';
 import { classes } from 'common/util';
