+++ src/core/renderers/Array1DRenderer/Array1DRenderer.module.scss
