+++ src/core/renderers/Array1DRenderer/index.js
