+++ src/core/renderers/Array2DRenderer/Array2DRenderer.module.scss
