+++ src/core/renderers/Array2DRenderer/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Array1DRenderer, Renderer } from 'core/renderers';
-import styles from './stylesheet.module.scss';
+import styles from './Array2DRenderer.module.scss';
 import { classes } from 'common/util';
 
 class Array2DRenderer extends Renderer {
