+++ src/core/renderers/ChartRenderer/ChartRenderer.module.scss
