+++ src/core/renderers/LogRenderer/LogRenderer.module.scss
